Fix tests and README
