+++ README.md
@@ -27,19 +27,19 @@ end
 Make API calls to Jellyfish:
 
 ```elixir
-client = Jellyfish.SDK.new("http://address-of-your-server.com")
+client = Jellyfish.Client.new("http://address-of-your-server.com")
 
 # Create room
-{:ok, room} = Jellyfish.SDK.Room.create_room(client, 10)  # pass maximum number of peers
+{:ok, room} = Jellyfish.Room.create(client, max_peers: 10)
 
 room.id
 # => "8878cd13-99a6-40d6-8d7e-8da23d803dab"
 
 # Add peer
-{:ok, peer} = Jellyfish.SDK.Peer.add_peer(client, room.id, "webrtc") # pass room id and type of peer
+{:ok, peer} = Jellyfish.Room.add_peer(client, room.id, "webrtc") # pass room id and type of peer
 
 # Delete peer
-:ok = Jellyfish.SDK.Peer.delete_peer(client, room.id, peer.id)
+:ok = Jellyfish.Room.delete_peer(client, room.id, peer.id)
 ```
 
 ## Copyright and License
